+++ api/animals/data_access/animal_viewing_alert_exhibit_name_provider.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from ...shared.enums.position import Position
+from ...types import Types
+
+
+class AnimalViewingAlertExhibitNameProvider():
+   @classmethod
+   def fetch_viewing_alert_exhibit_names(
+         cls,
+         conn: Types.Connection,
+         today: Types.DateKey ) -> list[ str ]:
+      cur = conn.cursor()
+
+      try:
+         data = cur.execute(
+            """   SELECT DISTINCT
+                     EXHIBIT
+                  FROM AnimalViewingAlert
+                  WHERE ALERT_END_DATE IS NULL
+                     OR ALERT_END_DATE >= ?
+                  ORDER BY EXHIBIT;
+            """,
+            ( today, ) )
+
+         return [ row[ Position.FIRST ] for row in data.fetchall() ]
+
+      finally:
+         cur.close()